==> Script/S8 - Capacity Clustering - Profiling.R <==
# Data Analytics for Education
# S8 - Capacity Clustering - Profiling
# July 19, 2015
# Author: Troy James R Palanca

# Libraries ---------------------------------------------------------------
library(dplyr)
library(reshape2)
library(ggplot2)
library(gridExtra)
library(ggmap)
library(extrafont)
loadfonts(device = "pdf")
loadfonts(device = "postscript")
library(rgl)
library(stringr)
library(xtable)

# Data --------------------------------------------------------------------
load("Data/D6 - Capacity Clustering.RData")

# Functions ---------------------------------------------------------------
ggplot_colors <- function (n) {
  colors = seq(15, 375, length = n+1)
  hcl(h = colors, l = 65, c = 100)[1:n]
}

# Visualizations of Final Clustering --------------------------------------

# Elementary
par3d(windowRect  = c(100, 100, 600, 600), family = "Open Sans", cex = 0.8)
plot3d(x = log10(schools_elem.dt$all.teacher.ratio),
       y = log10(schools_elem.dt$full.room.ratio),
       z = log10(schools_elem.dt$mooe.ratio),
       col = ggplot_colors(length(unique(schools_elem.dt$cluster)))
                                  [as.numeric(schools_elem.dt$cluster)],
       size = 1, box = F,
       xlab = "Teachers", ylab = "Rooms", zlab = "Budget")
legend3d("topright",
         legend = 1:length(unique(schools_elem.dt$cluster)),
         pch = 16, col = ggplot_colors(length(unique(schools_elem.dt$cluster))),
         cex=1, inset=c(0.02))
movie3d(spin3d(axis = c(0,0,1), rpm = 5), duration = 12,
        dir = paste(getwd(),"/Output", sep = ""),
        movie = "clusterplot")
file.rename(from = "Output/clusterplot.gif",
            to = "Output/O11 - Elementary School Clusters.gif")

# Secondary
par3d(windowRect  = c(100, 100, 600, 600), family = "Open Sans", cex = 0.8)
plot3d(x = log10(schools_seco.dt$all.teacher.ratio),
       y = log10(schools_seco.dt$full.room.ratio),
       z = log10(schools_seco.dt$mooe.ratio),
       col = ggplot_colors(length(unique(schools_seco.dt$cluster)))
       [as.numeric(schools_seco.dt$cluster)],
       size = 1, box = F,
       xlab = "Teachers", ylab = "Rooms", zlab = "Budget")
legend3d("topright",
         legend = 1:length(unique(schools_seco.dt$cluster)),
         pch = 16, col = ggplot_colors(length(unique(schools_seco.dt$cluster))),
         cex=1, inset=c(0.02))
movie3d(spin3d(axis = c(0,0,1), rpm = 5), duration = 12,
        dir = paste(getwd(),"/Output", sep = ""),
        movie = "clusterplot")
file.rename(from = "Output/clusterplot.gif",
            to = "Output/O13 - Secondary School Clusters.gif")

# Cluster Metric Charts ---------------------------------------------------
schools_elem_metrics.dt <- schools_elem.dt %>% select(cluster, all.teacher.ratio,
                                                      full.room.ratio, mooe.ratio) %>%
  melt(id.vars = c("cluster"), variable.name = "metric", value.name = "value")

png("Output/O12 - Elementary School Cluster Profiles.png", width = 600, height = 400)
ggplot(schools_elem_metrics.dt, aes(x = cluster, y = value, color = cluster)) +
  facet_wrap(~metric, ncol = 3, scales = "free_y") +
  scale_y_log10() +
  geom_boxplot(alpha = 0.05, outlier.size = 0) +
  theme_minimal() +
  ggtitle("Elementary Schools Cluster Profiles\n") +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.title.y = element_blank(),
        text = element_text(family = "Open Sans"),
        plot.title = element_text(hjust = 0, face = "bold", size = 16),
        legend.position = "none")
dev.off()

schools_seco_metrics.dt <- schools_seco.dt %>% select(cluster, all.teacher.ratio,
                                                      full.room.ratio, mooe.ratio) %>%
  melt(id.vars = c("cluster"), variable.name = "metric", value.name = "value")

png("Output/O14- Secondary School Cluster Profiles.png", width = 600, height = 400)
ggplot(schools_seco_metrics.dt, aes(x = cluster, y = value, color = cluster)) +
  facet_wrap(~metric, ncol = 3, scales = "free_y") +
  scale_y_log10() +
  geom_boxplot(alpha = 0.05, outlier.size = 0) +
  theme_minimal() +
  ggtitle("Secondary Schools Cluster Profiles\n") +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.title.y = element_blank(),
        text = element_text(family = "Open Sans"),
        plot.title = element_text(hjust = 0, face = "bold", size = 16),
        legend.position = "none")
dev.off()


# Cluster Profiling -------------------------------------------------------
# The data contained here is generated subjectively by studying the charts generated by prior code.

# Elementary Schools

schools_elem_profiles.dt <-
  schools_elem.dt %>% group_by(cluster) %>%
  summarise(teacher = 1/mean(all.teacher.ratio),
            rooms = 1/mean(full.room.ratio),
            mooe = mean(mooe.ratio))

schools_elem_profiles.dt$name <-
  c("Falling Behind",
    "Brain Drain",
    "Left Behind",
    "Head of the Pack",
    "Cash-Strapped",
    "Garden Variety")

schools_elem_profiles.dt$order <-
  c(2,3,1,6,5,4)

schools_elem_profiles.dt$description <-
  c("Although not the worst, these schools have started to fall behind, especially on budget.",
    "These schools risk empty rooms and unspent budgets due to lack of teachers.",
    "These schools are the most at-risk due to shortages in all three aspects: rooms, teachers, and budget. Rooms are especially in short supply.",
    "With ample numbers of teachers, rooms, and generous budgets, these schools are least likely to face capacity issues.",
    "Teachers and rooms may be enough, but these schools receive a lower budget per student than the rest.",
    "These are garden variety schools - not the best, but not the worst either.")

schools_elem_profiles_formatted.dt <-
  schools_elem_profiles.dt %>% arrange(desc(order)) %>%
  select(Cluster = name, Description = description, teacher, rooms, mooe,
         cluster.num = cluster) %>%
  mutate(Description = str_wrap(Description, 50)) %>%
  mutate(`Students\nper Teacher` = round(teacher,0),
         `Students\nper Room` = round(rooms, 0),
         `MOOE\nper Student` = round(mooe, 0)) %>%
  select(-teacher, -rooms, -mooe) %>%
  mutate(Cluster = factor(Cluster,
                          levels = c("Head of the Pack",
                                     "Cash-Strapped",
                                     "Garden Variety",
                                     "Brain Drain",
                                     "Falling Behind",
                                     "Left Behind")))

pdf("Output/O15A - Elementary Schools Cluster Profiles.pdf")
grid.table(schools_elem_profiles_formatted.dt %>% select(-cluster.num), core.just = "left",
           gpar.coretext = gpar(fontfamily = "Arial Narrow", fontsize = 10),
           gpar.coltext = gpar(fontfamily = "Arial Narrow", fontface = "bold",
                               fontsize = 12),
           show.rownames = F)
dev.off()

# Secondary Schools

schools_seco_profiles.dt <-
  schools_seco.dt %>% group_by(cluster) %>%
  summarise(teacher = 1/mean(all.teacher.ratio),
            rooms = 1/mean(full.room.ratio),
            mooe = mean(mooe.ratio))

schools_seco_profiles.dt$name <-
  c("Left Behind",
    "Close Second",
    "Garden Variety",
    "Brain Drain",
    "Jam Packed",
    "Head of the Pack")

schools_seco_profiles.dt$order <-
  c(1,5,2,4,3,6)

schools_seco_profiles.dt$description <-
  c("These schools are the most at-risk due to shortages in all three aspects: rooms, teachers, and budget.",
    "These schools aren't far from the top. Capacity is unlikely to be a pressing issue.",
    "These are garden variety schools - not the best, but not the worst either.",
    "These schools risk empty rooms and unspent budgets due to lack of teachers.",
    "Rooms are jam packed as facilities have failed to keep up with enrollment growth.",
    "With ample numbers of teachers, rooms, and generous budgets, these schools are least likely to face capacity issues.")

schools_seco_profiles_formatted.dt <-
  schools_seco_profiles.dt %>% arrange(desc(order)) %>%
  select(Cluster = name, Description = description, teacher, rooms, mooe,
         cluster.num = cluster) %>%
  mutate(Description = str_wrap(Description, 50)) %>%
  mutate(`Students\nper Teacher` = round(teacher,0),
         `Students\nper Room` = round(rooms, 0),
         `MOOE\nper Student` = round(mooe, 0)) %>%
  select(-teacher, -rooms, -mooe) %>%
  mutate(Cluster = factor(Cluster,
                          levels = c("Head of the Pack",
                                     "Close Second",
                                     "Brain Drain",
                                     "Jam Packed",
                                     "Garden Variety",
                                     "Left Behind")))

pdf("Output/O15B - Secondary Schools Cluster Profiles.pdf")
grid.table(schools_seco_profiles_formatted.dt  %>% select(-cluster.num), core.just = "left",
           gpar.coretext = gpar(fontfamily = "Arial Narrow", fontsize = 10),
           gpar.coltext = gpar(fontfamily = "Arial Narrow", fontface = "bold",
                               fontsize = 12),
           show.rownames = F)
dev.off()

# Write Out ---------------------------------------------------------------
rm(schools_elem_profiles.dt, schools_seco_profiles.dt,
   schools_elem_metrics.dt, schools_seco_metrics.dt, ggplot_colors)
save.image("Data/D7 - Cluster Profiles.RData")
